requirements.txt
streamlit
pandas
python-dotenv
openai
langchain
langchain-openai
google-generativeai